requirements.txt
discord.py==2.5.2
aiosqlite>=0.19.0
python-dotenv>=1.0.1
